Added a dummy search
