+++ client/src/pages/Home.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useEffect} from "react";
 import {useNavigate} from 'react-router-dom';
 import CustomSearchBar from '../components/CustomSearchBar';
 import '../assets/css/home.css';
@@ -6,9 +6,32 @@ import ReviewCardContainer from "../components/ReviewCardContainer";
 import Box from '@mui/material/Box';
 import AddIcon from '@mui/icons-material/Add';
 import IconButton from "@material-ui/core/IconButton";
+import axios from 'axios';
+import ReviewCard from '../components/ReviewCard';
+import {useGlobalState} from '../context/GlobalState';
 
 const Home = (props) => {
+    const [state, setState] = useGlobalState({reviewCards: []});
     const navigate = useNavigate();
+
+    useEffect(async () => {
+        await axios.get(`${window.location.origin.toString()}/review/getAllReviews`, {})
+        .then(res => {
+            console.log(res.data.data);
+            let allReviews = [];
+            res.data.data.forEach(x => allReviews.push(
+            <ReviewCard 
+                heading={`${x['User']['UserLogin']}, ${(new Date(x['ReviwedAt'])).toLocaleString('default', { month: 'short', day: 'numeric', year: 'numeric'})} at ${(new Date(x['ReviwedAt'])).toLocaleString('default', { timeStyle: 'long'})}`} 
+                brandName={x['Product']['ProductBrand']} 
+                productName={x['Product']['ProductName']} 
+                review={x['ReviewText']}
+            />));
+            setState({...state, allReviewCards: allReviews});
+        })
+        .catch(error => {
+            console.log (`Error fetching all the reviews while mounting the home page with error: ${error}`);
+        })
+    }, [])
 
     // Currently we are fetching all the data there exists for MVP. 
     // TODO: Modify this to include paged fetching
